the lottie heap changed from SRAM_DTC_cm7 to SRAM_OC1, large images saved in BOARD_SDRAM (LV_ATTRIBUTE_LARGE_CONST in lv_conf.h), smooth animations

+++ AirconGUI/sdk/Core/drivers/fsl_mipi_dsi.c
@@ -873,7 +873,7 @@ void DSI_WriteApbTxPayloadExt(
         payloadSizeLocal += 1U;
     }
 
-    assert(payloadSizeLocal <= (uint16_t)FSL_DSI_TX_MAX_PAYLOAD_BYTE);
+    assert(payloadSizeLocal <= FSL_DSI_TX_MAX_PAYLOAD_BYTE);
 
     /* The first 4-byte. */
     if (sendDscCmd)
